Ownership deletion secret (#21)
Secrets are owned by GitopsClusters

This changes the ownership of Secrets to be owned by the referencing GitopsCluster.

This does two things, the Secret is owned by the Cluster which causes cascading removal. It also annotates the Cluster to indicate that the cluster-controller should not wait for the Secret to be removed.

Trigger reconciliation on change to the GitopsCluster.